requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.12
click==8.1.7
cohere==4.27
colorama==0.4.6
coloredlogs==15.0.1
exceptiongroup==1.1.3
fastapi==0.99.1
fastavro==1.8.2
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.1
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
numpy==1.26.0
onnxruntime==1.16.0
openai==0.28.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pydantic==1.10.12
pydantic_core==2.6.3
PyJWT==2.8.0
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.0
tqdm==4.66.1
twilio==8.8.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
